Add custom HiCache routing and benchmark notebook
