+++ python/sglang/srt/server_args.py
@@ -437,6 +437,7 @@ class ServerArgs:
 
     # Hierarchical cache
     enable_hierarchical_cache: bool = False
+    hicache_impl: str = "vanilla"
     hicache_ratio: float = 2.0
     hicache_size: int = 0
     hicache_write_policy: str = "write_through"
@@ -3210,6 +3211,13 @@ class ServerArgs:
             "--enable-hierarchical-cache",
             action="store_true",
             help="Enable hierarchical cache",
+        )
+        parser.add_argument(
+            "--hicache-impl",
+            type=str,
+            choices=["vanilla", "custom"],
+            default=ServerArgs.hicache_impl,
+            help="Choose the HiCache implementation to use when hierarchical cache is enabled.",
         )
         parser.add_argument(
             "--hicache-ratio",
